pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drheaderplus-mcp"
version = "0.1.0"
description = "MCP server for DrHeaderPlus security header auditing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "drheaderplus>=3.0.3",
    "mcp[cli]>=1.26",
]

[project.scripts]
drheaderplus-mcp = "drheaderplus_mcp.server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.anyio]
backends = ["asyncio"]
